LBFGS refined grid search done and tested.
Most significant changes were to gpacquisition.py where I had to
squeeze the 0th axis of the gradient when the provided X included
a single test point.

Two tests were added: the first test_global_solve_branin() checks
that global_solve() can optimize a known function (Branin). As a
side note, for this I had to compute and code up the gradient for
Branin, thought it'd be good to know in case you ever need it. The
second test test_global_solve_direct() checks that after a few
observations global_solve() and solve_direct() produce the same
optimizer of EI.