Adding kwarg max for policies needing maximization
Broke things for now until Matt, fixes `solve_direct()` to take in
the same argument.

Right now the argument is hardcoded as `max=True` in
`gppolicy.get_next()` but should really be a self variable that is
specific to a policy. E.g. improvement methods would be maximized
while LCB would be minimized.