Simplify solve_bayesopt.
This forgets about being able to select noise/kernel/etc. This goes back to an
approach we've used before where you pass an entire gp model. If none is given
then a default is used (matern3 with default hyperparameters) and a default
prior. These defaults could be tightened up a bit though.